+++ src/math/my_fmin.c
@@ -5,7 +5,7 @@
 ** Get the lowest integer.
 */
 
-int my_fmin(int x, int y)
+double my_fmin(double x, double y)
 {
     return x < y ? x : y;
 }
